Updated Back-End branch

+++ src/main/java/me/caravanweb/resources/api/CaravanasResource.java
@@ -1,11 +1,11 @@
 package me.caravanweb.resources.api;
 
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,12 +13,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 import me.caravanweb.profiles.Caravanas;
 import me.caravanweb.services.CaravanasService;
+import me.caravanweb.services.UsuarioService;
 
 @RestController
 @CrossOrigin
@@ -26,6 +27,8 @@ import me.caravanweb.services.CaravanasService;
 public class CaravanasResource {
 	@Autowired
 	private CaravanasService service;
+	@Autowired
+	private UsuarioService serviceu;
 	
 	@GetMapping
 	public ResponseEntity<List<Caravanas>> findAll() { 
@@ -44,24 +47,25 @@ public class CaravanasResource {
 		return ResponseEntity.ok().body(obj);
 	}
 	
-	@PostMapping
+	@PostMapping(value = "/register", consumes = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseStatus(HttpStatus.CREATED)
-	public ResponseEntity<Caravanas> create(@RequestBody Caravanas c){
-		service.add(c);
-		URI uri = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
-				.buildAndExpand(c.getId()).toUri();
-		return ResponseEntity.created(uri).body(c);
+	public @ResponseBody ResponseEntity<String> create(@RequestBody Caravanas c){
+		String body =  service.add(c);
+		return ResponseEntity.ok().body(body);
 	}
 
     @PostMapping(value = "/delete/{id}")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<String> delete(@PathVariable Integer id){
-    	Caravanas caravana = service.findById(id);
     	String body = "";
-		boolean status = service.delete(caravana);
-		if (status) {
-			body = "Deletado.";
-		}
+    	Caravanas caravana = service.findById(id);
+    	if (caravana != null) {
+    		body = serviceu.cleanCaravanaUsers(id);
+    		boolean status = service.delete(caravana);
+    		if (status) {
+    			body = "Deletado.";
+    		}
+    	}
 		return ResponseEntity.ok().body(body);    
 	}
